Don't attempt to remove old-hooks directory if non-empty
Attempt to delete the old-hooks directory _only_ if it is empty,
otherwise we get an exception.

Travis has been passing this test for a while, but it fails when running
on macOS, so there is likely some platform difference at play here.

Fixes #554

+++ lib/overcommit/installer.rb
@@ -138,8 +138,8 @@ module Overcommit
           FileUtils.mv(hook_file, old_hooks_path)
         end
       end
-      # Remove old-hooks directory if empty
-      FileUtils.rmdir(old_hooks_path)
+      # Remove old-hooks directory if empty (i.e. no old hooks were preserved)
+      FileUtils.rmdir(old_hooks_path) if Dir.entries(old_hooks_path).size <= 2
     end
 
     def restore_old_hooks
